AspAIra/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
boto3==1.26.137
streamlit==1.22.0
requests==2.28.2
python-dotenv==0.19.2
pyyaml==6.0
pytest==7.3.1
aiohttp==3.8.1 